+++ src/dftbp/dftbplus/main.F90
@@ -1724,7 +1724,8 @@ contains
         call writeCurrentGeometry(this%geoOutFile, this%pCoord0Out, this%tLatOpt, this%tMd,&
             & this%tAppendGeo.and.iGeoStep>0, this%tFracCoord, this%tPeriodic, this%tHelical,&
             & this%tPrintMulliken, this%species0, this%speciesName, this%latVec, this%origin,&
-            & iGeoStep, iLatGeoStep, this%nSpin, this%qOutput, this%velocities)
+            & iGeoStep, iLatGeoStep, this%nSpin, this%qOutput, this%velocities, this%coord,&
+            & this%extendedGeomFile, this%species)
       endif
     end if
 
@@ -2033,7 +2034,7 @@ contains
         call writeCurrentGeometry(this%geoOutFile, this%pCoord0Out, .false., .true., .true.,&
             & this%tFracCoord, this%tPeriodic, this%tHelical, this%tPrintMulliken, this%species0,&
             & this%speciesName, this%latVec, this%origin, iGeoStep, iLatGeoStep, this%nSpin,&
-            & this%qOutput, this%velocities)
+            & this%qOutput, this%velocities, this%coord, this%extendedGeomFile, this%species)
       end if
       this%coord0(:,:) = this%newCoords
       if (this%tWriteDetailedOut  .and. this%deltaDftb%nDeterminant() == 1) then
@@ -2134,7 +2135,7 @@ contains
 
   !> Does the operations that are necessary after a lattice vector update
   subroutine handleLatticeChange(latVecs, sccCalc, tblite, tStress, extPressure, mCutOff,&
-      & repulsive, dispersion, solvation, cm5Cont, recVecs, recVecs2p, cellVol, recCellVol,&
+      & repulsive, dispersion, solvation, cm5Cont, recVecs, invLatVecs, cellVol, recCellVol,&
       & extLatDerivs, cellVecs, rCellVecs, boundaryCond)
 
     !> Lattice vectors
@@ -2171,7 +2172,7 @@ contains
     real(dp), intent(out) :: recVecs(:,:)
 
     !> Reciprocal lattice vectors in units of 2 pi
-    real(dp), intent(out) :: recVecs2p(:,:)
+    real(dp), intent(out) :: invLatVecs(:,:)
 
     !> Unit cell volume
     real(dp), intent(out) :: cellVol
@@ -2191,12 +2192,8 @@ contains
     !> Boundary conditions on the calculation
     type(TBoundaryConds), intent(in) :: boundaryCond
 
-    cellVol = abs(determinant33(latVecs))
-    recVecs2p(:,:) = latVecs
-    call invert33(recVecs2p)
-    recVecs2p = transpose(recVecs2p)
-    recVecs = 2.0_dp * pi * recVecs2p
-    recCellVol = abs(determinant33(recVecs))
+    call boundaryCond%handleBoundaryChanges(latVecs, invLatVecs, recVecs, cellVol, recCellVol)
+
     if (tStress) then
       call derivDeterminant33(extLatDerivs, latVecs)
       extLatDerivs(:,:) = extPressure * extLatDerivs
@@ -2224,7 +2221,7 @@ contains
        call cm5Cont%updateLatVecs(latVecs)
        mCutoff = max(mCutOff, cm5Cont%getRCutOff())
     end if
-    call getCellTranslations(cellVecs, rCellVecs, latVecs, recVecs2p, mCutOff)
+    call getCellTranslations(cellVecs, rCellVecs, latVecs, invLatVecs, mCutOff, boundaryCond)
 
   end subroutine handleLatticeChange