+++ src/dftbp/math/lapackroutines.F90
@@ -300,12 +300,12 @@ contains
     !> Pivot indices, row i of the matrix was interchanged with row ipiv(i).
     integer, intent(out) :: ipiv(:)
 
-    !> Number of rows of the matrix to decomposea. (Necessary if different from the number of rows
+    !> Number of rows of the matrix to decompose. (Necessary if different from the number of rows
     !> of the passed matrix)
     integer, optional, intent(in) :: nRow
 
-    !> Number of rows of the matrix to decompose. (Necessary if different from the number of columns
-    !> of the passed matrix)
+    !> Number of columns of the matrix to decompose. (Necessary if different from the number of
+    !> columns of the passed matrix)
     integer, optional, intent(in) :: nColumn
 
     !> Error flag. Zero on successful exit. If not present, any lapack error causes program
@@ -359,12 +359,12 @@ contains
     !> Pivot indices, row i of the matrix was interchanged with row ipiv(i).
     integer, intent(out) :: ipiv(:)
 
-    !> Number of rows of the matrix to decomposea. (Necessary if different from the number of rows
-    !> of the passed matrix)
+    !> Number of rows of the matrix to decompose. (Necessary if different from the number of rows of
+    !> the passed matrix)
     integer, optional, intent(in) :: nRow
 
-    !> Number of rows of the matrix to decompose. (Necessary if different from the number of columns
-    !> of the passed matrix)
+    !> Number of columns of the matrix to decompose. (Necessary if different from the number of
+    !> columns of the passed matrix)
     integer, optional, intent(in) :: nColumn
 
     !> Error flag. Zero on successful exit. If not present, any lapack error causes program
@@ -418,12 +418,12 @@ contains
     !> Pivot indices, row i of the matrix was interchanged with row ipiv(i).
     integer, intent(out) :: ipiv(:)
 
-    !> Number of rows of the matrix to decomposea. (Necessary if different from the number of rows
-    !> of the passed matrix)
+    !> Number of columns of the matrix to be decomposed. (Necessary if different from the number of
+    !> rows of the passed matrix)
     integer, optional, intent(in) :: nRow
 
-    !> Number of rows of the matrix to decompose. (Necessary if different from the number of columns
-    !> of the passed matrix)
+    !> Number of columns of the matrix to decompose. (Necessary if different from the number of
+    !> columns of the passed matrix)
     integer, optional, intent(in) :: nColumn
 
     !> Error flag. Zero on successful exit. If not present, any lapack error causes program
@@ -477,12 +477,12 @@ contains
     !> Pivot indices, row i of the matrix was interchanged with row ipiv(i).
     integer, intent(out) :: ipiv(:)
 
-    !> Number of rows of the matrix to decomposea. (Necessary if different from the number of rows
+    !> Number of rows of the matrix to decompose. (Necessary if different from the number of rows
     !> of the passed matrix)
     integer, optional, intent(in) :: nRow
 
-    !> Number of rows of the matrix to decompose. (Necessary if different from the number of columns
-    !> of the passed matrix)
+    !> Number of columns of the matrix to decompose. (Necessary if different from the number of
+    !> columns of the passed matrix)
     integer, optional, intent(in) :: nColumn
 
     !> Error flag. Zero on successful exit. If not present, any lapack error causes program